pyproject.toml
[tool.poetry]
name = "졸프스터디"
version = "0.1.0"
description = ""
authors = ["YEWON\\kuty2 <kuty2004@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.55.0"
langchain = "^1.2.10"
langchain-openai = "^1.1.10"
langchain-community = "^0.4.1"
langchain-postgres = "^0.0.17"
psycopg = "^3.3.3"
pdfplumber = "^0.11.9"
python-dotenv = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
